Moved get_risk_models inside risklib [skip hazardlib]
