+++ openquake/commonlib/tests/riskmodels_test.py
@@ -24,7 +24,8 @@ import numpy
 from numpy.testing import assert_almost_equal
 from openquake.baselib.general import writetmp
 from openquake.hazardlib import InvalidFile, nrml
-from openquake.commonlib import nrml_examples, riskmodels
+from openquake.commonlib import nrml_examples
+from openquake.risklib import riskmodels
 from openquake.qa_tests_data.scenario_damage import case_4b
 
 EXAMPLES_DIR = os.path.dirname(nrml_examples.__file__)
